+++ src/pages/SignUp/SignUp.tsx
@@ -5,28 +5,14 @@ import { SignInType, SignUpType } from '../../types/auth.ts'
 import AuthInput from '../../components/common/AuthInput/AuthInput.tsx'
 import { VALID_RULES } from '../../constants/validRules.ts'
 import { theme } from '../../styles/theme.ts'
-import SelectButton from '../../components/common/SelectButton/SelectButton.tsx'
-import { useEffect, useState } from 'react'
-import { abilities, gender, playStyle } from '../../constants/signUpSelect.ts'
+import { useEffect } from 'react'
 import BasicButton from '../../components/common/BasicButton/BasicButton.tsx'
 import { useDuplicate } from '../../hooks/useDuplicate.ts'
+import UserInfoSelect from '../../components/common/UserInfoSelect/UserInfoSelect.tsx'
+import { useSelect } from '../../hooks/useSelect.ts'
 
-interface SelectedInfo {
-  gender: string
-  playStyle: string
-  abilities: string
-}
-// TODO: 중복검사 통과 후 결과값을 담을 state를 생성하고 에러메시지 자리에 뿌려주기
+// TODO: 중복검사 통과 후 결과값을 담을 state 를 생성하고 에러메시지 자리에 뿌려주기
 export default function SignUp() {
-  const [genderSelected, setGenderSelected] = useState<number>(0)
-  const [styleSelected, setStyleSelected] = useState<number>(0)
-  const [abilitiesSelected, setAbilitiesSelected] = useState<number>(0)
-  const [select, setSelect] = useState<SelectedInfo>({
-    gender: '남자',
-    playStyle: '공격적',
-    abilities: '벨러스',
-  })
-
   const {
     handleSubmit,
     control,
@@ -44,40 +30,27 @@ export default function SignUp() {
     },
   })
 
+  const selectedValue = useSelect()
+
   useEffect(() => {
     if (Object.keys(errors).length > 0) {
       window.scrollTo({ top: 0, left: 0, behavior: 'smooth' })
     }
   }, [errors])
 
-  const passowrd = watch('password')
+  const password = watch('password')
 
   const { emailPassed, idPassed, nicknamePassed } = useDuplicate()
 
   const handleSignUp = (sigInData: SignInType) => {
     console.log(sigInData)
-    console.log(select)
+    console.log(selectedValue.select)
 
     if (!emailPassed && !idPassed && !nicknamePassed) {
       alert('모든 중복 검사를 완료해 주세요.')
     }
 
     // TODO: 서버에 보내기전 모든 항목 공백제거 후 보낼것
-  }
-
-  const handleGenderSelect = (index: number) => {
-    setGenderSelected(index)
-    setSelect({ ...select, gender: gender[index] })
-  }
-
-  const handleStyleSelect = (index: number) => {
-    setStyleSelected(index)
-    setSelect({ ...select, playStyle: playStyle[index] })
-  }
-
-  const handleAbilitiesSelect = (index: number) => {
-    setAbilitiesSelected(index)
-    setSelect({ ...select, abilities: abilities[index] })
   }
 
   const handleIdDuplicate = (id: string) => {
@@ -181,7 +154,7 @@ export default function SignUp() {
               name={'passwordConfirm'}
               control={control}
               placeholder={'비밀번호를 한번 더 입력해 주세요.'}
-              rules={VALID_RULES.PASSWORD_CONFIRM(passowrd)}
+              rules={VALID_RULES.PASSWORD_CONFIRM(password)}
             />
           </S.InputWrapper>
           <S.InputWrapper>
@@ -241,62 +214,9 @@ export default function SignUp() {
               rules={VALID_RULES.BIRTHDAY}
             />
           </S.InputWrapper>
-          <S.SelectSection>
-            <>
-              <S.InputLabel>성별</S.InputLabel>
-              <S.SelectWrapper>
-                {gender.map((gender, index) => (
-                  <div key={gender}>
-                    <SelectButton
-                      type={'button'}
-                      $bgColor={theme.colors.white}
-                      $width={'100%'}
-                      onClick={() => handleGenderSelect(index)}
-                      $isSelected={genderSelected === index}
-                    >
-                      {gender}
-                    </SelectButton>
-                  </div>
-                ))}
-              </S.SelectWrapper>
-            </>
-            <>
-              <S.InputLabel>플레이 스타일</S.InputLabel>
-              <S.SelectWrapper>
-                {playStyle.map((style, index) => (
-                  <div key={style}>
-                    <SelectButton
-                      type={'button'}
-                      $bgColor={theme.colors.white}
-                      $width={'100%'}
-                      onClick={() => handleStyleSelect(index)}
-                      $isSelected={styleSelected === index}
-                    >
-                      {style}
-                    </SelectButton>
-                  </div>
-                ))}
-              </S.SelectWrapper>
-            </>
-            <>
-              <S.InputLabel>자신있는 능력</S.InputLabel>
-              <S.SelectWrapper>
-                {abilities.map((ability, index) => (
-                  <div key={ability}>
-                    <SelectButton
-                      type={'button'}
-                      $bgColor={theme.colors.white}
-                      $width={'100%'}
-                      onClick={() => handleAbilitiesSelect(index)}
-                      $isSelected={abilitiesSelected === index}
-                    >
-                      {ability}
-                    </SelectButton>
-                  </div>
-                ))}
-              </S.SelectWrapper>
-            </>
-          </S.SelectSection>
+
+          <UserInfoSelect selected={selectedValue} />
+
           <BasicButton
             type={'submit'}
             $bgColor={theme.colors.blue}
